feat: 代码优化，新增异常处理

+++ apps/message-router-server/src/app/events/events.gateway.ts
@@ -13,12 +13,22 @@ export class EventsGateway implements OnGatewayConnection, OnGatewayDisconnect {
 
   handleConnection(client: Socket): any {
     const device = client?.request?.headers?.device as string;
+    if (!device) {
+      this.logger.log('未获取到设备代码，中断!');
+      client.disconnect(true);
+      return;
+    }
+    this.logger.log(`客户端 ${device} 成功连接!`);
     this.bridgeService.addDevices(device, client);
   }
 
   handleDisconnect(client: Socket): void {
-    this.logger.log('disconnect');
     const device = client?.request?.headers?.device as string;
+    if (!device) {
+      this.logger.log('非法设备断开!')
+      return;
+    }
+    this.logger.log(`客户端 ${device} 断开连接!`);
     this.bridgeService.removeDevices(device);
   }
